refactor: move AppointmentDetailsScreen logic to its own file
